Il timeout funziona. Il giocare contro il PC funziona. Da commentare e controllarne correttezza e coerenza.

+++ src/TriClient.c
@@ -16,13 +16,14 @@ void printError(const char *);
 void init_data();
 void print_board();
 void move();
+void pc_move();
 void signal_handler(int);
 void removeIPCs();
 void remove_pid_from_game();
 void set_sig_handlers();
 void remove_terminal_echo();
 void restore_terminal_echo();
-void p(int, int);
+int p(int, int);
 void v(int, int);
 
 // Attributi del terminale
@@ -44,17 +45,40 @@ int semaphores;
 // Indirizzo di memoria condivisa che contiene la matrice di gioco.
 char *board = NULL;
 
+// Istante dell'ultima pressione di Ctrl+C
+int sigint_timestamp = 0;
+
 // Indice nell'array info->client_pid del giocatore
 int player;
+
+// Indica se il client ha giocato una mossa (serve a gestire il Ctrl+C durante la partita).
+int move_played = 0;
+
+pid_t child;
 
 // Set di segnali ricevibili dal processo.
 sigset_t processSet;
 
 int main(int argc, char *argv[]){
 
+    child = -1;
+    srand(time(NULL));
+
+    /** TODO: Limita inserimento argomenti! 
+     *        Metti i nomi a Giocatore e Computer.
+    */
+    if(argc > 1){
+        if(argv[1][0] == '*' && argv[1][1] == '\0'){
+            child = fork();
+            if(child == -1)
+                printError(NO_CHILD_CREATED_ERR);
+        }
+    }
+
     set_sig_handlers();
 
-    printf("%s", CLEAR);
+    if(child != 0)
+        printf("%s", CLEAR);
 
     tcgetattr(STDIN_FILENO, &termios);
 
@@ -64,12 +88,22 @@ int main(int argc, char *argv[]){
     if(kill(server, SIGUSR1) == -1)
         printError(SIGUSR1_SEND_ERR);
 
-    printf("%s\n", WAITING);
+    if(child != 0)
+        printf("%s\n", WAITING);
 
     // Si aspetta di essere in due!
-    pause();
+    /** TODO: Controllo su pressione Ctrl+C in attesa. */
+    int started;
+    do {
+        pause();
 
-    printf("\n%s\n", GAME_STARTING);
+        p(INFO_SEM, NOINT);
+        started = info->game_started;
+        v(INFO_SEM, NOINT);
+    } while (!started);
+
+    if(child != 0)
+        printf("\n%s\n", GAME_STARTING);
 
     p(INFO_SEM, NOINT);
 
@@ -85,21 +119,25 @@ int main(int argc, char *argv[]){
     v(INFO_SEM, NOINT);
 
     // Stampa la matrice vuota
-    print_board();
-    
-    remove_terminal_echo();
+    if(child != 0){
+        print_board();
+        remove_terminal_echo();
+    }
 
     int my_semaphore = (player == 0) ? CLIENT1_SEM : CLIENT2_SEM;
 
     while(partitaInCorso) {
-        p(my_semaphore, WITHINT);
+        while(p(my_semaphore, WITHINT) == -1);
 
         // Svuota il buffer del terminale e ignora tutti i caratteri inseriti.
-        tcflush(STDIN_FILENO, TCIFLUSH);
-        restore_terminal_echo();
+        if(child != 0){
+            tcflush(STDIN_FILENO, TCIFLUSH);
+            restore_terminal_echo();
+        }
 
         // In ogni caso si stampa lo stato della partita
-        print_board();
+        if(child != 0)
+            print_board();
 
         p(INFO_SEM, NOINT);
         // Il server comunica se la partita è terminata o meno
@@ -108,20 +146,33 @@ int main(int argc, char *argv[]){
 
         if(partitaInCorso){
             // La partita non è finita. Si procede.
-            move();
-            print_board();
+            if(child != 0){
+                /** FAQ: Se non si mette il countdown questo while è necessario ??? */
+                do {
+                    move();
+                } while(move_played == 0);
+            } else {
+                pc_move();
+            }
 
-            remove_terminal_echo();
+            move_played = 0;
+
+            if(child != 0){
+                print_board();
+                remove_terminal_echo();
+            }
 
             v(SERVER, WITHINT);
         } else {
-            printf("\n%s", GAME_ENDED);
-            if(info->winner == getpid())
-                printf(" %s\n\n", YOU_WON);
-            else if(info->winner == info->server_pid)
-                printf(" %s\n\n", DRAW);
-            else
-                printf(" %s\n\n", YOU_LOST);
+            if(child != 0){
+                printf("\n%s", GAME_ENDED);
+                if(info->winner == getpid())
+                    printf(" %s\n\n", YOU_WON);
+                else if(info->winner == info->server_pid)
+                    printf(" %s\n\n", DRAW);
+                else
+                    printf(" %s\n\n", YOU_LOST);
+            }
         }
     }
 
@@ -138,8 +189,9 @@ int main(int argc, char *argv[]){
  * Imposta gli handler dei segnali da catturare.
 */
 void set_sig_handlers(){
-    if(signal(SIGINT, signal_handler) == SIG_ERR)
-        printError(SIGINT_HANDLER_ERR);
+    if(child != 0)
+        if(signal(SIGINT, signal_handler) == SIG_ERR)
+            printError(SIGINT_HANDLER_ERR);
 
     if(signal(SIGTERM, signal_handler) == SIG_ERR)
         printError(SIGTERM_HANDLER_ERR);
@@ -149,6 +201,12 @@ void set_sig_handlers(){
 
     if(signal(SIGHUP, signal_handler) == SIG_ERR)
         printError(SIGHUP_HANDLER_ERR);
+
+    /*
+    if(signal(SIGALRM, signal_handler) == SIG_ERR)
+        printError(SIGALRM_HANDLER_ERR);
+    */
+    
 }
 
 /**
@@ -182,7 +240,7 @@ void restore_terminal_echo(){
  * @param: semnum - il semaforo su cui eseguire p
  * @param: no_int - dice se si abilita la cattura di segnali durante l'attesa su semaforo.
 */
-void p(int semnum, int no_int){
+int p(int semnum, int no_int){
 
     // Disabilita la cattura di tutti i segnali (catturabili)
     if(no_int){
@@ -205,8 +263,14 @@ void p(int semnum, int no_int){
      * il flusso d'esecuzione successivo alla p.
     */
 
-    if(semop(semaphores, &p, 1) == -1)
-        printError(P_ERR);
+    int code;
+    if((code = semop(info->semaphores, &p, 1)) == -1){
+        // Vero errore solo se non si riceve EINTR ( = si è ricevuto un segnale)
+        if(errno != EINTR)
+            printError(P_ERR);
+    }
+
+    return code;
 }
 
 /**
@@ -261,20 +325,45 @@ void print_board(){
     }
 
     printf(" %s%s  '   '  \n\n", FIELD_TAB, BOARD_TAB);
+    printf("Per abbandonare, premere due volte Ctrl+C in %d secondi.\n", MAX_SECONDS);
 }
 
 /**
  * Esegue una mossa. Si suppone che ad inserimento errato equivalga concedere il turno ???.
- * TODO: Migliore acquisizione dati (interfaccia).
+ * TODO: Dopo Ctrl+C si vedono i malanni.
 */
 void move(){
     char coord[4] = {0};
-    char output[31];
-    snprintf(output, 31, "> Inserisci una coordinata %c: ", info->signs[player]);
-    write(STDOUT_FILENO, output, 31);
-    read(STDIN_FILENO, coord, 3);
+    char output[51] = {0};
 
-    if(coord[2] != '\n' || (coord[1] < '1' && coord[1] > '3') || 
+    int seconds = info->timeout;
+    int bytesRead = -1;
+
+    printf("\nTempo a disposizione: %d secondi.\n", seconds);
+    
+    snprintf(output, 50, "\r> Inserisci una coordinata %c: ", info->signs[player]);
+    write(STDOUT_FILENO, output, 50);
+
+    /** NB:*/
+    // Fa in modo che l'alarm faccia chiudere la read(). Se in delta funziona senza meglio, sennò AMEN.
+    struct sigaction act;
+    act.sa_flags = ~SA_RESTART;
+    act.sa_handler = signal_handler;
+    sigaction(SIGALRM, &act, NULL);
+    alarm(seconds);
+
+    bytesRead = read(STDIN_FILENO, coord, 4);
+
+    move_played = 1;
+
+    if(bytesRead <= 0){
+        info->move_made[0] = 'N';
+        info->move_made[1] = 'V';
+        info->move_made[2] = '\0';
+        return;
+    }
+
+    if(coord[2] != '\n' || !(coord[1] >= '1' && coord[1] <= '3') || 
         !((coord[0] >= 'a' && coord[0] <= 'c') || (coord[0] >= 'A' && coord[0] <= 'C'))){
             info->move_made[0] = 'N';
             info->move_made[1] = 'V';
@@ -301,7 +390,28 @@ void move(){
 }
 
 /**
+ * Esegue una mossa casuale generata dal Computer.
+*/
+void pc_move(){
+
+    int riga, colonna;
+    
+    do {
+        riga = rand() % 3;
+        colonna = rand() % 3;
+    } while (board[(riga * 3) + colonna] != ' ');
+
+    board[(riga * 3) + colonna] = info->signs[player];
+
+    info->move_made[0] = (char) ('a' + riga);
+    info->move_made[1] = (char) ('1' + colonna);
+    info->move_made[2] = '\0';
+
+}
+
+/**
  * Ottiene i dati inizializzati dal server riguardo la partita da giocare.
+ * TODO: Metti bene i log a terminale del Computer.
 */
 void init_data(){
     key_t lobbyShmKey = ftok(PATH_TO_FILE, FTOK_KEY);
@@ -398,37 +508,56 @@ void removeIPCs(){
 void signal_handler(int sig){
     if(sig == SIGINT || sig == SIGHUP){
 
-        // Ritorna indietro per scrivere sopra al carattere ^C
-        printf("\r");
-        printf("%s\n", BLANK_LINE);
-        printf("%s\n\n", QUITTING);
-        
-        // Si notifica al server che si vuole abbandonare la partita dopo aver rimosso il client
-        // dalle info di gioco e rimosso gli IPC.
-        remove_pid_from_game();
-        removeIPCs();
+        if(child != 0)
+            printf("\r");
 
-        if(kill(server, SIGUSR2) == -1)
-            printError(SIGUSR2_SEND_ERR);
+        int now = time(NULL);
+        if(now - sigint_timestamp < MAX_SECONDS || sig == SIGHUP) {
 
-        restore_terminal_echo();
+            if(child != 0){
+                printf("%s\n", BLANK_LINE);
+                printf("%s\n\n", QUITTING);
+            }
+            
+            // Si notifica al server che si vuole abbandonare la partita dopo aver rimosso il client
+            // dalle info di gioco e rimosso gli IPC.
+            remove_pid_from_game();
+            removeIPCs();
 
-        exit(0);
+            if(kill(server, SIGUSR2) == -1)
+                printError(SIGUSR2_SEND_ERR);
+
+            if(child != 0)
+                restore_terminal_echo();
+
+            exit(0);
+
+        } else {
+            sigint_timestamp = now;
+            // Ritorna indietro per scrivere sopra al carattere ^C
+            printf("  \033[A\n");
+        }
 
     } else if(sig == SIGTERM){
         // Terminazione causata dal server.
-        printf("\r%s\n", BLANK_LINE);
 
-        // P e V non necessarie: si è sicuri che info->winner ha già il valore che deve assumere.
-        if(info->winner == info->server_pid)
-            printf("%s\n\n", SERVER_STOPPED_GAME);
-        else
-            printf("%s\n\n", GAME_WON);
+        if(child != 0){
+            printf("\r%s\n", BLANK_LINE);
+
+            // P e V non necessarie: si è sicuri che info->winner ha già il valore che deve assumere.
+            if(info->winner == info->server_pid)
+                printf("%s\n\n", SERVER_STOPPED_GAME);
+            else
+                printf("%s\n\n", GAME_WON);
+        }
 
         removeIPCs();
 
-        restore_terminal_echo();
+        if(child != 0)
+            restore_terminal_echo();
 
         exit(0);
+    } else if(sig == SIGALRM){
+        
     }
 }
